nutrition-video-analysis/deploy/requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.0.3
python-multipart==0.0.6

# Deep Learning & Computer Vision
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4

# SAM2 Requirements
hydra-core==1.3.2
iopath==0.1.10
tqdm==4.66.1

# Metric3D Requirements
mmengine==0.10.2
mmcv-lite==2.1.0
timm==0.9.12

# RAG & NLP
sentence-transformers==2.2.2
faiss-cpu==1.7.4  # Use faiss-gpu for GPU acceleration
PyPDF2==3.0.1
pandas==2.1.3
openpyxl==3.1.2  # For Excel file reading

# External APIs
google-generativeai==0.3.1

# Database
psycopg2-binary==2.9.9  # Optional: for PostgreSQL support

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging & Monitoring (optional)
# prometheus-client==0.19.0
# sentry-sdk==1.39.1

# Development Tools (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0

